Make fsm_init, fsm_enter, fsm_exit, fsm_process_event functions return simply an INT.
It was very confusing for them to return a new event, because it wasn't clear when and how
this new event should be processed: before entering new state? After that? This was also
making fsm_process_event a recursive function very difficult to follow in extreme cases.

So now the only way to generate new events is to enqueue them at the end of the FIFO or to manually call fsm_process_event.